pyproject.toml
[project]
name = "cdm-tools-mcp-server"
version = "0.1.0"
description = "MCP Server exposing Audit CDM Agent tools for ERP-to-CDM data transformation"
readme = "README.md"
requires-python = ">=3.11"

dependencies = [
    "fastapi>=0.115.12",
    "fastmcp>=2.12.5",
    "mcp[cli]>=1.14.0",
    "uvicorn>=0.34.2",
    "databricks-sdk>=0.63.0",
    "pydantic>=2",
    # CDM tool dependencies
    "pandas>=2.2",
    "openpyxl>=3.1",
    "chardet>=5.0",
]

[dependency-groups]
dev = [
    "ruff>=0.8.0",
    "databricks-mcp",
    "pytest",
    "requests",
]

[project.scripts]
cdm-mcp-server = "server.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["server", "cdm_tools"]

[tool.ruff]
line-length = 100
target-version = "py311"
